Fix compilation of transport/SessionHandle.h
Currently SessionHandle, SessionHolder, and Session must be defined
in order. Any other order will not compile.

Merge these headers in order to satisfy the following objectives:

1) All headers compile in isolation
2) Header order inclusion does not matter

This is the simplest way to fix the problem, and seems to be the only
way to fix it without changing the class definitions.
